+++ .circleci/config.yml
@@ -204,8 +204,7 @@ jobs:
             python -m pip install wheel
             python -m pip install -r requirements.txt
             pip install git+https://github.com/Deci-AI/super-gradients.git@${CIRCLE_BRANCH}
-            python src/super_gradients/examples/train_from_recipe_example/train_from_recipe.py --config-name=cifar10_resnet experiment_name=cifar10_resnet_sanity training_hyperparams.max_epochs=3
-            python -m unittest /tests/recipe_training_tests/recipe_sanity_unit_test.py
+            python /tests/recipe_training_tests/recipe_sanity_test.py --config-name=cifar10_resnet experiment_name=cifar10_resnet_sanity training_hyperparams.max_epochs=3 +goal_metric_val=0.9
 
 
 workflows:
