+++ src/super_gradients/recipes/cifar10_resnet.yaml
@@ -24,7 +24,8 @@ resume: False
 training_hyperparams:
   resume: ${resume}
 
-
+multi_gpu: DDP
+num_gpus: 4
 ckpt_root_dir:
 
 architecture: resnet18_cifar
